[Phase 3] Implement TmuxBackend for external tmux multiplexer support
Created complete TmuxBackend implementation:
- src/terminal/tmux_backend.rs: Full tmux backend (386 lines)
- All 11 TerminalBackend trait methods implemented
- Uses external tmux for session multiplexing

Features:
- Session naming: kimichat-{pid}-{session_id} (avoids conflicts)
- Checks tmux availability at initialization
- Creates sessions with clean environment
- Maps string session IDs to tmux session names

Implementation details:
- launch_session: Uses 'tmux new-session -d' with size and working dir
- send_keys: Direct 'tmux send-keys' passthrough
- get_screen: Uses 'tmux capture-pane -p' with optional color codes
- list_sessions: Queries tmux for session metadata
- kill_session: Kills tmux session and removes tracking
- get_cursor_position: Reads #{cursor_y},#{cursor_x} from tmux
- resize_session: Uses 'tmux resize-window'
- get_scrollback: Captures history with -S flag
- set_scrollback: Sets history-limit option
- capture_start: Uses 'tmux pipe-pane' for logging
- capture_stop: Stops pipe-pane and returns file stats
- session_exists: Checks 'tmux has-session'

Manager integration:
- Updated TerminalManager::with_backend() to instantiate TmuxBackend
- Graceful fallback to PTY if tmux initialization fails
- Backend name reporting for user visibility

Fixes:
- Fixed lifetime issue in resolve_terminal_backend() (env var handling)
- Fixed partial move issue in main.rs (use ref in pattern match)
- Fixed type mismatch in pty_backend.rs (u64 to usize cast)

Status: Compiles successfully with no errors. Both backends fully operational.

+++ src/terminal/pty_backend.rs
@@ -176,7 +176,7 @@ impl TerminalBackend for PtyBackend {
         let mut session = session.lock().unwrap();
         let (path, bytes, duration) = session.stop_capture()?;
         self.capture_files.remove(session_id);
-        Ok((path.display().to_string(), bytes, duration))
+        Ok((path.display().to_string(), bytes as usize, duration))
     }
 
     async fn session_exists(&self, session_id: &str) -> bool {
